Moved configuration test clases to their own package, made private the @Bean PropertySourcesPlaceholderConfigurer.
