+++ src/test/java/embl/ebi/variation/eva/pipeline/steps/GenesLoadTest.java
@@ -20,19 +20,15 @@ import embl.ebi.variation.eva.VariantJobsArgs;
 import embl.ebi.variation.eva.pipeline.gene.GeneFilterProcessor;
 import embl.ebi.variation.eva.pipeline.gene.GeneLineMapper;
 import embl.ebi.variation.eva.pipeline.gene.FeatureCoordinates;
-import embl.ebi.variation.eva.pipeline.jobs.AnnotationConfig;
-import embl.ebi.variation.eva.pipeline.jobs.InitDBConfig;
+import embl.ebi.variation.eva.pipeline.config.InitDBConfig;
 import embl.ebi.variation.eva.pipeline.jobs.JobTestUtils;
 import embl.ebi.variation.eva.pipeline.steps.readers.GeneReader;
 import embl.ebi.variation.eva.pipeline.steps.writers.GeneWriter;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.opencb.opencga.storage.core.variant.VariantStorageManager;
 import org.springframework.batch.item.ExecutionContext;
 import org.springframework.batch.item.ItemProcessor;
-import org.springframework.batch.item.ItemWriter;
-import org.springframework.batch.item.file.FlatFileItemReader;
 import org.springframework.batch.test.MetaDataInstanceFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.context.ContextConfiguration;
