+++ src/test/java/embl/ebi/variation/eva/pipeline/steps/VariantsAnnotLoadTest.java
@@ -19,6 +19,7 @@ import com.mongodb.*;
 import embl.ebi.variation.eva.VariantJobsArgs;
 import embl.ebi.variation.eva.pipeline.MongoDBHelper;
 import embl.ebi.variation.eva.pipeline.annotation.load.VariantAnnotationLineMapper;
+import embl.ebi.variation.eva.pipeline.config.AnnotationConfig;
 import embl.ebi.variation.eva.pipeline.jobs.*;
 import embl.ebi.variation.eva.pipeline.steps.readers.VariantAnnotationReader;
 import embl.ebi.variation.eva.pipeline.steps.writers.VariantAnnotationWriter;
@@ -28,14 +29,11 @@ import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.opencb.biodata.models.variant.annotation.VariantAnnotation;
-import org.opencb.opencga.storage.core.variant.VariantStorageManager;
 import org.opencb.opencga.storage.mongodb.variant.DBObjectToVariantAnnotationConverter;
 import org.springframework.batch.core.BatchStatus;
 import org.springframework.batch.core.ExitStatus;
 import org.springframework.batch.core.JobExecution;
 import org.springframework.batch.item.ExecutionContext;
-import org.springframework.batch.item.ItemWriter;
-import org.springframework.batch.item.file.FlatFileItemReader;
 import org.springframework.batch.item.file.FlatFileParseException;
 import org.springframework.batch.test.JobLauncherTestUtils;
 import org.springframework.batch.test.MetaDataInstanceFactory;
